Extracted StrategyManager class from SearchMain
